backend/requirements.txt
fastapi
uvicorn
pymongo
psycopg2-binary
python-dotenv
